refactor: phase 2 dedup — kiwoom, discovery, notifications, data_sources
#1369 — Kiwoom consolidation:
- Move kiwoom/*.py driver files → brokers/kiwoom/ (single home for all kiwoom code)
- kiwoom/ becomes a backward-compat shim package (per-module shims + __init__ shim)
- brokers/kiwoom/__init__.py: exports all driver symbols (adapter excluded to avoid circular imports)
- Update internal cross-module imports to use brokers.kiwoom.*

#1370 — Notification deprecation:
- portfolio/notifier.py: add DeprecationWarning pointing to notification_dispatcher
- Canonical path is api.services.notification_dispatcher for all new code

#1371 — data_sources/ skeleton:
- Create data_sources/{market,technical}/ package structure
- data_sources/market/ohlcv.py: canonical home for get_ohlcv and related fetch utilities
- utils/fetch.py: converted to backward-compat shim → data_sources.market.ohlcv

#1372 — discovery/ → screener/:
- Move discovery/indicators.py → screener/indicators.py
- Move discovery/decision.py → screener/decision.py
- Move discovery/evaluator.py → screener/evaluator.py
- Move discovery/evaluators/ → screener/evaluators/
- discovery/ becomes a shim package with DeprecationWarning
- screener/conditions/accumulation.py: update from discovery.indicators → screener.indicators

Test results: 48 failed / 2556 passed (vs baseline 51 failed / 2553 passed — zero regressions, +3 pass)
