+++ data_sources/market/__init__.py
@@ -1,0 +1,2 @@
+"""data_sources.market — OHLCV and market data fetchers."""
+from data_sources.market.ohlcv import get_ohlcv  # noqa: F401
