+++ kiwoom/screen_manager.py
@@ -1,9 +1,2 @@
-"""Compatibility wrapper for screen number allocator.
-
-This module keeps the phase-document path (`kiwoom/screen_manager.py`) while
-delegating to the active implementation in `kiwoom/realtime.py`.
-"""
-
-from kiwoom.realtime import ScreenManager
-
-__all__ = ["ScreenManager"]
+"""kiwoom.screen_manager — backward-compatibility shim → brokers.kiwoom.screen_manager."""
+from brokers.kiwoom.screen_manager import *  # noqa: F401, F403
