+++ kiwoom/tr_request.py
@@ -1,9 +1,2 @@
-"""Compatibility wrapper for TR request helpers.
-
-This module keeps the phase-document path (`kiwoom/tr_request.py`) while
-delegating to the active implementation in `kiwoom/tr.py`.
-"""
-
-from kiwoom.tr import KiwoomTrClient, TrRequest
-
-__all__ = ["KiwoomTrClient", "TrRequest"]
+"""kiwoom.tr_request — backward-compatibility shim → brokers.kiwoom.tr_request."""
+from brokers.kiwoom.tr_request import *  # noqa: F401, F403
